2007-03-29  Sebastien Pouliot  <sebastien@ximian.com> 	* tiffcodec.c: Fix TIFF crashes on SPARC (it's amazing what an integer 	can do to get promoted).
svn path=/trunk/libgdiplus/; revision=75168

+++ src/tiffcodec.c
@@ -215,7 +215,7 @@ gdip_load_tiff_properties(TIFF *tiff, BitmapData *bitmap_data)
 	uint16	samples_per_pixel;
 	uint16	bits_per_sample;
 	uint16	planar_configuration;
-	uint16	image_length;
+	uint32	image_length;
 	uint16	strips_per_image;
 	uint32	rows_per_strip;
 	uint32	tile_length;
